changed content length from int to long

+++ netty/src/main/java/xingu/netty/http/HttpResponseBuilder.java
@@ -74,7 +74,7 @@ public class HttpResponseBuilder
         return this;
     }
 
-    public HttpResponseBuilder withContentLength(int length)
+    public HttpResponseBuilder withContentLength(long length)
     {
         return withHeader(CONTENT_LENGTH, length);
     }
